+++ src/components/inputs/ToolInput.vue
@@ -137,7 +137,7 @@ export default {
 							let code = '';
 							if (this.controlTools) {
 								this.tools.forEach(function(tool) {
-									if (tool.heaters.length) {
+									if (tool && tool.heaters.length) {
 										const temps = tool.heaters.map(() => this.inputValue, this).join(':');
 										code += `G10 P${tool.number} ${this.active ? 'S' : 'R'}${temps}\n`;
 									}