+++ src/components/panels/JobEstimationsPanel.vue
@@ -4,7 +4,7 @@
 			<v-icon small class="mr-1">mdi-clock</v-icon> {{ $t('panel.jobEstimations.caption') }}
 		</v-card-title>
 
-		<v-card-text class="text-center pb-1">
+		<v-card-text class="text-center pb-2">
 			<v-row dense>
 				<v-col class="d-flex flex-column">
 					<strong>
@@ -38,7 +38,7 @@
 						{{ $t('panel.jobEstimations.slicer') }}
 					</strong>
 					<span>
-						{{ $displayTime(slicerTimeleft) }}
+						{{ $displayTime(slicerTimeLeft) }}
 					</span>
 				</v-col>